Replace nw.js with Electron.js
Electron seems to have better documentation and a more active
development cycle. And it allows us to get rid of the npm usage. No need
to install node-modules anymore; instead we need to download the
Electron installers. Given that it looks that there is a way to ship
multi-platform installations, this is not that bad.

The code is copied/inspired from:

    https://github.com/bchazalet/scalajs-electron-skeleton/tree/8530fc6c027acc645b4cfdf633357f40daac6589