Handle multi-line text changes in UI
This is not the best solution, but the best solution that seems to be
possible right now. Nvim doesn't support the transfer of text change
events yet. The API is there, but no functionality is yet implemented to
send text change events. Therefore we always get the entire file content
after each change and refresh the entire UI. This is the way how it is
done in a lot of other Nvim clients and we have to wait for a more
efficient way to update the UI in future.