Sync active Nvim windows between client and server
This requires that we always have to send the active window in the
protocol. Another important change is that we need to ignore the next
WinEnter event that comes from Nvim whenever the selection changed on
the client. The client doesn't need an update in this case, since it
already has all the information. Furthermore we don't have to think
about what happens if the client update occurs too early, i.e. we didn't
yet set the new selection.