Do no longer depend on akka-stream TCP
It worked but we don't need the overhead of this implementation. It is
enough to only depend on the Msgpack library.

Furthermore, this fixes a bug when multiple responses are passed to the
client in one flow. Previously we only read the first response and
discarded the rest, now we continue reading until no further data is
found.