Introduce new algorithm for window tree creator
It took me now forever to come up with this algorithm but it was worth
it. It is a lot simpler than the previous one, even though it has more
line of codes in size. It passes more tests than the previous one but
there is still one failing test left.

Grouping similar elements together is a key design point of functional
programs and I'm happy that it could be used here. The previous
recursive algorithm just got too complex. Maybe it is possible to
simplify the current algorithm even further once the last test passed.