Handle annotation declarations and type aliases
Actually, I wanted to fix annotation declarations but it turned out that
I also had to fix type aliases. We want to index type aliases because
from there we can easily follow to the underlying declarations.

Annotations can leak in an implicit defined Annotation type, which we
want to ignore.