Add logging for Jena
We need a log4j configuration for Jena. I have no idea if it is possible
to get rid of it and to use only one configuration for both Akka and
Jena. It is not trivial to implement because Jena is not in the actor
system and therefore needs to log with %c instead of %X{akkaSource}. See
the link to PatternLayout for documentation about the characters.