Wait for artifact entry to be fully indexed
So far we have added entries to be indexed to the actor message queue,
which meant that the message queue was huge and blocked other requests.
The new solution is much better but has the disadvantage that we need an
ask call. I guess an even better solution would be to use a reactive
stream that can only process one item at a time but for such a solution
larger refactorings would be necessary.