Intorduce RequestActor
With this actor it should be easier in the long term to propagate errors
through the actor system. The problem is that once when an error
happened, the REST request needs a corresponding response. So far errors
were caught internally and the request timed out. It is still not
perfect because the indexer for example still needs to answer with a Try
but one day I may find a way to get rid of this one too.