Index new SPARQL update queries instead of JSONLD schemas
These JSONLD schemas are simply not something I want to use. They are
hard to read and complicated to handle by a machine. The SPARQL queries
are a better way to index the schema information.

We know also can index subtype relationships, which leads to the problem
that multiple owners of a declaration are found as one can see by the
changed tests. This is a problem because actually only one (the least
upper bound one) owner should be found. Instead of one owner it would
also make sense to order the owners somehow. But for now I don't know
how to make it better. I need to revisit this in the future.