Add linked data service handling framework
Finally, it is done! The first version of my very own service handling
framework is implemented. It was an incredible amount of work needed to
get this far but it feels great. It is based completely on linked data,
which means that requests to the framework are in linked data, that the
communication between services is in linked data and that responses from
the services are also linked data. And SPARQL queries are needed to
access the information that is sent over the wire. I really like it.

Now I only need to get this thing production ready...

But before I go there I can now start to use some services from a
client. Like find usages or semantic highlighting that is shown on
Github - I'm slowly getting there.