Introduce sbt projects for services
Each service can have library dependencies. In order to avoid clashes
between services, they need to be separated in their own projects.

We start by moving the scalac converter tests to its own project.
Furthermore the tests are moved to a project that does not have a
dependency to the Scala compiler. This makes it possible to later reuse
the tests in dotc and scalac.