Replace N3 files by Turtle files
Instead of using the Notation3 syntax, we want use the Turtle syntax.
They are basically the same, but N3 supports some extensions which we
probably never need, therefore it seems to be better to limit the files
to Turtle syntax only. In fact, the N3 extensions are not supported by
Jena and probably not by any oher library out there, therefore it makes
even less sense to use it.