Move scalac tests back to ScalacConverterTest
These tests were moved to ScalaCompilerTest because they can be shared
with DotcConverterTest. However, the dotty converter is not yet
implemented. It makes more sense to move tests to ScalaCompilerTest once
the test is fixed in the dotty converter. This way we can see which
changes in the converter fix a test.