Get rid of version information in URIs
So far we don't need the version information in order to disambiguate
the triples. I'm not even sure if we ever need this information. Since
everything is supposed to be linked data it should be theoretically
possible to just stay compatible when changes happen in the ontologies.
I have to find out in future, for now it is just overhead to keep the
version information.