Include generated constructors of Scala classes to index
The generated constructors and its parameter are semantic information
and therefore it should be able to find them through queries.

The only thing that is skipped right now are the super calls within
generated constructors, which call the default constructor of
java.lang.Object. While this would also be semantic information I'm
simply not sure if we want highlight this fact in the index. We can
always add them if they are needed one day.

A lot of implicit constructors and parameters are now found, which makes
problems if a constructor has multiple parameters. They all have the
same position and therefore they overwrite each other in the index. We
have to add the parameter position or something else to the generated
ids otherwise we can't disambiguate them.