Add SourceFile attachment to package declarations
A package declaration like `pakage pkg` doesn't have an associated file
because only symbols within a class have associated files. This is an
annoying limitation in the compiler but we can't do anything against it.
We still need SourceFile attachments for package declarations, the fix
therefore is to associate it with the file of the compilation unit.

We only have to ensure to set the position before the association code
block is called, otherwise we don't know what the difference between a
package declaration in a Scala source file and a package declaration of
a bytecode reference is.

The problem only happens in the compiler plugin and not in the unit
tests because in the latter the SourceFile attachments are hardcoded.