Index ref to decl for imports correctly
So far, the import selectors didn't have the correct ref to decl
attached. Since the selectors do not have a symbol we had to look it up
manually. Furthermore, since top level imports are not part of a class,
they don't have an associated file (only classes have), we had to
introduce a new method `mkDeclWithPos` that can handle this fact.