Add tests for refs that access the this reference
I'm not completely sure if the `this` should be indexed when it is
written explicitly. Since it is a keyword the only thing that could be
done by indexing it is to jump to the class definition, which is not the
most important behavior therefore I don't change anything right now.