Index type annotations at parameter correctly
We want to index all explicitly written type annotations and if the type
is inferred, the type shall be indexed with an offset position that
points to the declaration.

In case of primary constructors, we want to index the type ascription
twice. Once for the class member and another time for the constructor
parameter.