Read config file on startup by passing it as parameter
The default config file that is loaded is now ~/amora.conf. Its
fall-back is the src/main/resources/application.conf file. A config file
can be passed as a parameter on startup of the server. The actor system
uses the combined configs if necessary instead of just the
application.conf.